requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
wandb>=0.15.0
pyyaml>=6.0
pandas>=1.5.0
numpy>=1.24.0
tqdm>=4.65.0
accelerate>=0.20.0
bitsandbytes>=0.41.0